\name{prepare_figure2a_data}
\alias{prepare_figure2a_data}
\title{Prepare the data needed to reproduce Figure 2A}
\usage{prepare_figure2a_data(xlsx_file, sheet_name = "Sheet1", skip = 2, top_n = 1246, reverse_rows = TRUE)}
\description{Reads the supplementary Excel file, filters the data and computes a z-score matrix for plotting.}
